# contract/openapi.yaml
openapi: 3.0.3
info:
  title: BuryMe API
  version: 1.0.0
  description: |
    Backend contract for **BuryMe** — a peer-to-peer IOU / micro-lending tracker.
    Derived directly from PRD v2.7, §5 (Conceptual Model), §6–7 (Requirements &
    Scenarios), §8 (Entity Attribute Reference), §11 (Dependencies & Integrations),
    and §12 (Validation & Error Handling).

    ## Architecture
    - **Thin clients.** All business logic, persistence, and third-party integration
      live in the Express/Node backend (§11.0). React (web) and Flutter (mobile)
      consume this API only — neither has direct DB or third-party access.
    - **Auth is Firebase-backed** (§11.1.2). Phone/email OTP and credential
      management happen entirely client-side via the Firebase Auth SDK. The client
      sends the resulting Firebase ID token as `Authorization: Bearer <token>` on
      every request; Express verifies it via the Firebase Admin SDK. There is
      therefore no `/login`, `/send-otp`, or `/verify-otp` endpoint here — those
      never touch this API.
    - **First login provisions a BuryMe user.** A valid Firebase token with no
      matching BuryMe profile yields `404` on `GET /auth/me`; the client then calls
      `POST /auth/register`.

    ## Request vs. Obligation — the core modeling decision
    Per §5 and §8.5/§8.6, **Request** and **Obligation** are two distinct persisted
    resources, not two states of one resource:
...
      It never re-enters a negotiation state.

...
    pre-filled.
...
servers:
...
security:
  - firebaseIdToken: []
paths:
...
  /users/search:
    get:
      tags: [Users]
      summary: Search registered, verified users (§6.1.3)
      description: >
        Supports lookup by display name, phone, or email (§7.1.3). Returns only
        verified, registered accounts. Min 2 / max 100 characters (§12.2.1).
      operationId: searchUsers
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 100 }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Matching users.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserSummary" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get a user's public profile summary
      operationId: getUser
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Public profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserSummary" }
        "404": { $ref: "#/components/responses/NotFound" }
  /users/me:
    patch:
      tags: [Users]
      summary: Update own profile (§6.1.2)
      description: Display name and/or profile photo. §12.2.1 constraints apply.
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileRequest" }
      responses:
        "200":
          description: Updated profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /users/me/telebirr:
    get:
      tags: [Users]
      summary: Get own Telebirr verification status (§8.4)
      operationId: getTelebirr
      responses:
        "200":
          description: Telebirr account state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TelebirrAccount" }
    post:
      tags: [Users]
      summary: Register / re-verify a Telebirr number (§6.1.6)
      description: >
        Prompted contextually, not at registration. Updating the number resets
        status to Unverified until re-verified (§7.1.6). §12.2.1 format constraints
        apply.
      operationId: verifyTelebirr
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyTelebirrRequest" }
      responses:
        "200":
          description: >
            Updated Telebirr account state (Unverified while verification is in
            progress). Call `POST /users/me/telebirr/otp` next to start OTP
            verification of the saved number.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TelebirrAccount" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /users/me/telebirr/otp:
    post:
      tags: [Users]
      summary: Send (or resend) a verification code to the saved Telebirr number
      description: >
        Sends a fresh 6-digit one-time code by SMS to the `telebirr_number`
        recorded via `POST /users/me/telebirr`. Codes expire after 5 minutes;
        a new code may be requested at most once per 60 seconds and replaces
        any earlier one. Ownership of a Telebirr number is proven by receiving
        this SMS — Chapa's integration does not verify account ownership.
      operationId: sendTelebirrOtp
      responses:
        "200":
          description: Code sent.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TelebirrOtpChallenge" }
        "409":
          description: No Telebirr number is saved yet, or the account is already Verified (`STATUS_CONFLICT`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: A code was sent less than 60 seconds ago (`DUPLICATE_SUBMISSION`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /users/me/telebirr/verify:
    post:
      tags: [Users]
      summary: Confirm the Telebirr verification code
      description: >
        Checks the code against the pending challenge. On success the account
        becomes `Verified` and `verified_at` is set. Five wrong attempts
        invalidate the challenge; request a new code with
        `POST /users/me/telebirr/otp`.
      operationId: confirmTelebirrOtp
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfirmTelebirrOtpRequest" }
      responses:
        "200":
          description: Telebirr account now `Verified`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TelebirrAccount" }
        "400":
          description: Wrong or expired code (`VALIDATION_ERROR`, field `code`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: No verification is pending, attempts are exhausted, or the account is already Verified (`STATUS_CONFLICT`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ══════════════════════════════════════════════════════════════
  # REQUESTS — unified Borrow / Lend / Repayment (§8.5)
  # ══════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Firebase ID token (1-hour expiry, §11.1.2). Verified on the backend via the
        Firebase Admin SDK on every protected route.
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    ValidationError:
      description: One or more input fields failed validation (§12.3.2 `VALIDATION_ERROR`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: VALIDATION_ERROR
              message: The due date must be a future date.
              field: proposed_due_date
...
    NotFound:
      description: Record does not exist or is not accessible to this user (§12.3.2 `NOT_FOUND`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    ErrorCode:
      type: string
      description: Standard error codes per §12.3.2.
      enum:
        - VALIDATION_ERROR
        - DUPLICATE_SUBMISSION
        - STATUS_CONFLICT
        - BALANCE_EXCEEDED
        - CONCURRENT_MODIFICATION
        - PAYMENT_GATEWAY_UNAVAILABLE
        - PAYMENT_GATEWAY_ERROR
        - RECIPIENT_UNVERIFIED
        - PARTIAL_WRITE_ROLLED_BACK
        - UNAUTHENTICATED
        - UNAUTHORIZED
        - NOT_FOUND
        - INTERNAL_ERROR
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: Plain-language, user-safe. Never exposes stack traces, DB ids, or raw third-party payloads (§12.1).
            field:
              type: string
              nullable: true
              description: Present only for field-level validation errors; omitted for system errors (§12.3.1).

    # ── Enums (verbatim from §8) ─────────────────────────────────
...
    TelebirrStatus:
      type: string
      enum: [Unverified, Verified]
...
    UpdateProfileRequest:
      type: object
      properties:
        display_name:
          type: string
          minLength: 2
          maxLength: 50
          pattern: "^[A-Za-z0-9 _-]+$"
        profile_photo_url:
          type: string
          nullable: true
          description: Set via a separate photo-upload flow (multer/S3, §11.1.6); this field references the resulting URL.
      additionalProperties: false
...
    VerifyTelebirrRequest:
      type: object
      required: [telebirr_number]
      properties:
        telebirr_number:
          type: string
          pattern: '^(\+2519\d{8}|09\d{8})$'
          description: §12.2.1 Ethiopian Telebirr format.
      additionalProperties: false
    TelebirrOtpChallenge:
      type: object
      description: State of the verification code just sent by `POST /users/me/telebirr/otp`. The code itself is never returned.
      required: [expires_at, resend_available_at]
      properties:
        expires_at:
          type: string
          format: date-time
          description: When the code stops being accepted (5 minutes after sending).
        resend_available_at:
          type: string
          format: date-time
          description: Earliest time a new code may be requested (60 seconds after sending).
    ConfirmTelebirrOtpRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          pattern: '^\d{6}$'
          description: The 6-digit code received by SMS.
      additionalProperties: false
    TelebirrAccount:
      type: object
      required: [account_id, telebirr_number, verification_status, verified_at, created_at]
      properties:
        account_id: { type: string }
        telebirr_number: { type: string, nullable: true }
        verification_status: { $ref: "#/components/schemas/TelebirrStatus" }
        verified_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    UserSummary:
      type: object
      required: [user_id, display_name, profile_photo_url]
      properties:
        user_id: { type: string }
        display_name: { type: string }
        profile_photo_url: { type: string, nullable: true }
    User:
      allOf:
        - $ref: "#/components/schemas/UserSummary"
        - type: object
          required:
            [identifier, verification_status, telebirr, notification_preferences, created_at]
          properties:
            identifier:
              type: string
              description: Email or phone used for login/search — sourced from the verified Firebase identity.
            verification_status: { type: string, enum: [Verified] }
            telebirr:
              allOf: [{ $ref: "#/components/schemas/TelebirrAccount" }]
              nullable: true
              description: Null until the user completes Telebirr verification (§6.1.3) — not set at registration.
            notification_preferences:
              allOf: [{ $ref: "#/components/schemas/NotificationPreferences" }]
              description: >
                Current push switches. Returned here rather than from a
                dedicated GET so the preferences screen can render its state
                from the session's own profile read.
            created_at: { type: string, format: date-time }

    # ── Requests (§8.5) ───────────────────────────────────────────
...
    NotificationPreferences:
      type: object
      description: >
        Per-category push switches (§7.1.5, §6.1.5). Each maps to a family of
        §13.2 catalog events. These control push delivery only — every
        notification is still written to the in-app feed whatever is set
        here. All fields are optional on PATCH; omitted ones are left alone.
      required: [requests, group_expenses, payments, schedules, settlements, reminders, telebirr]
      properties:
        requests:
          type: boolean
          description: New borrow requests, counter-proposals, and lending records sent to you (BR-*, RR-01, LR-*).
        group_expenses:
          type: boolean
          description: Shares assigned to you, acknowledgements, and consolidated expense updates (GE-*).
        payments:
          type: boolean
          description: Chapa payments, external records waiting on your confirmation, and confirmed payments (PAY-*).
        schedules:
          type: boolean
          description: Schedules proposed to you, accepted schedules, and installment changes (RS-01 through RS-03).
        settlements:
          type: boolean
          description: Settlement suggestions, responses needed, and completed settlements (BSS-*).
        reminders:
          type: boolean
          description: Upcoming and overdue installments, and payments still waiting on a response (RS-04 through RS-06, REM-01).
        telebirr:
          type: boolean
          description: Prompts to verify Telebirr, and when someone you owe verifies theirs (TB-*).